名称変更 + Refactoring
