+++ server/streamer/heya/handler.go
@@ -1,75 +1,116 @@
 package heya
 
 import (
-	"errors"
+	"context"
+	"time"
 
 	"github.com/gofrs/uuid"
+	"github.com/hackathon-21winter-05/HiQidas/model"
 	"github.com/hackathon-21winter-05/HiQidas/server/protobuf/ws"
-	"google.golang.org/protobuf/proto"
+	"github.com/hackathon-21winter-05/HiQidas/service/utils"
+	"google.golang.org/protobuf/types/known/wrapperspb"
 )
 
-func (hs *HeyaStreamer) heyaWSHandler(mes *heyaCliMessage) error {
-	var WsHeyaData *ws.WsHeyaData
-
-	err := proto.Unmarshal(mes.body, WsHeyaData)
-	if err != nil {
-		_ = hs.sendErrorMes(mes.clientID, err.Error())
-		return err
-	}
-
-	switch WsHeyaData.GetPayload().(type) {
-	case *ws.WsHeyaData_CreateHiqidashi:
-		err := hs.createHiqidashiHandler(mes.userID, mes.heyaid, WsHeyaData.GetCreateHiqidashi())
-		if err != nil {
-			_ = hs.sendErrorMes(mes.clientID, err.Error())
-			return err
-		}
-		return nil
-
-	case *ws.WsHeyaData_EditHiqidashi:
-		err := hs.editHiqidashiHandler(mes.userID, mes.heyaid, WsHeyaData.GetEditHiqidashi())
-		if err != nil {
-			_ = hs.sendErrorMes(mes.clientID, err.Error())
-			return err
-		}
-		return nil
-
-	case *ws.WsHeyaData_DeleteHiqidashi:
-		err := hs.deleteHiqidashiHandler(mes.heyaid, WsHeyaData.GetDeleteHiqidashi())
-		if err != nil {
-			_ = hs.sendErrorMes(mes.clientID, err.Error())
-			return err
-		}
-		return nil
-
-	default:
-		_ = hs.sendErrorMes(mes.clientID, "unknown payload")
-		return errors.New("unknown payload")
-	}
-}
-
-func (hs *HeyaStreamer) sendHeyaMes(heyaID uuid.UUID, mes *ws.WsHeyaData) error {
-	buffer, err := proto.Marshal(mes)
+func (hs *HeyaStreamer) createHiqidashiHandler(userID, heyaID uuid.UUID, body *ws.WsCreateHiqidashi) error {
+	parentUUID, err := uuid.FromString(body.GetParentId())
 	if err != nil {
 		return err
 	}
 
-	hs.sendToHeya(heyaID, buffer)
-	return nil
-}
+	created, err := hs.ser.CreateHiqidashi(context.Background(), userID, heyaID, parentUUID)
+	if err != nil {
+		return err
+	}
 
-func (hs *HeyaStreamer) sendErrorMes(clientID uuid.UUID, message string) error {
-	mes, err := proto.Marshal(
+	var drawing *wrapperspb.StringValue = nil
+	if created.Drawing.Valid {
+		drawing = &wrapperspb.StringValue{Value: created.Drawing.String}
+	}
+
+	res := &ws.WsSendHiqidashi{
+		Hiqidashi: &ws.Hiqidashi{
+			Id:          created.ID.String(),
+			CreatorId:   created.CreatorID.String(),
+			ParentId:    &wrapperspb.StringValue{Value: created.ParentID.UUID.String()},
+			Title:       created.Title,
+			Description: created.Description,
+			Drawing:     drawing,
+			ColorCode:   created.ColorCode,
+		},
+	}
+
+	err = hs.sendHeyaMes(heyaID,
 		&ws.WsHeyaData{
-			Payload: &ws.WsHeyaData_Error{
-				Error: &ws.WsError{Message: message},
+			Payload: &ws.WsHeyaData_SendHiqidashi{
+				SendHiqidashi: res,
 			},
 		})
 	if err != nil {
 		return err
 	}
 
-	hs.sendToClient(clientID, mes)
+	return nil
+}
+
+func (hs *HeyaStreamer) editHiqidashiHandler(userID, heyaID uuid.UUID, body *ws.WsEditHiqidashi) error {
+	uuid, err := uuid.FromString(body.GetId())
+	if err != nil {
+		return err
+	}
+
+	hiqidashi := &model.NullHiqidashi{}
+	hiqidashi.ID = uuid
+	if body.GetTitle() != nil {
+		hiqidashi.Title = utils.NullStringFrom(body.GetTitle().Value)
+	}
+	hiqidashi.Description = utils.NullString()
+	if body.GetDrawing() != nil {
+		hiqidashi.Drawing = utils.NullStringFrom(body.GetDrawing().Value)
+	}
+	if body.GetColorCode() != nil {
+		hiqidashi.ColorCode = utils.NullStringFrom(body.GetColorCode().Value)
+	}
+	hiqidashi.LastEditorID = userID
+	hiqidashi.UpdatedAt = time.Now()
+
+	err = hs.ser.UpdateHiqidashiByID(context.Background(), hiqidashi)
+	if err != nil {
+		return err
+	}
+
+	err = hs.sendHeyaMes(heyaID,
+		&ws.WsHeyaData{
+			Payload: &ws.WsHeyaData_EditHiqidashi{
+				EditHiqidashi: body,
+			},
+		})
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (hs *HeyaStreamer) deleteHiqidashiHandler(heyaID uuid.UUID, body *ws.WsDeleteHiqidashi) error {
+	uuid, err := uuid.FromString(body.GetId())
+	if err != nil {
+		return err
+	}
+
+	err = hs.ser.DeleteHiqidashiByID(context.Background(), uuid)
+	if err != nil {
+		return err
+	}
+
+	err = hs.sendHeyaMes(heyaID,
+		&ws.WsHeyaData{
+			Payload: &ws.WsHeyaData_DeleteHiqidashi{
+				DeleteHiqidashi: body,
+			},
+		})
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
